requirements.txt
sentence-transformers
faiss-cpu
langchain-text-splitters
python-docx
numpy
flask
flask-cors
